Module Tezos_benchmark_type_inference_021_PsQuebec

module Inference : sig ... end

module Int_map : sig ... end

module Mikhailsky : sig ... end

Mikhailsky: Michelson in Micheline form, with typed holes and annotations. Mikhailsky terms are hash-consed.

module Mikhailsky_prim : sig ... end

Mikhailsky primitives correspond to Michelson primitives plus special "holes" for instructions and data.

module Monads : sig ... end

module Stores : sig ... end

module Type : sig ... end

Michelson types, hash-consed.

module Uf : sig ... end